requirements.txt
charset-normalizer==2.0.12
certifi==2021.10.8
idna==3.3
urllib3==1.26.9
python-pptx==0.6.21
XlsxWriter==3.0.3
requests==2.27.1
lxml==4.9.1
Pillow==9.2.0
streamlit
numpy
altair
pandas
scipy
click
streamlit-cookies-manager
extra-streamlit-components